Bot_with_Dots/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
tldextract==5.1.2
python-dotenv==1.0.1
joblib==1.4.2
openai==1.40.3
httpx==0.27.0
python-multipart==0.0.9
